requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
idna==3.4
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
validators==0.20.0
